backend/requirements.txt
# ========================
# Core Backend
# ========================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ========================
# LangChain + Gemini
# ========================
langchain==0.3.27
langchain-core==0.3.79
langchain-community==0.3.31
langchain-google-genai==2.1.12

# ========================
# Embeddings (HuggingFace - Free Alternative)
# ========================
sentence-transformers==2.2.2
torch==2.0.1

# ========================
# Vector Store
# ========================
faiss-cpu==1.12.0

# ========================
# Document Processing
# ========================
pypdf==4.0.0

# ========================
# Validation / Data Models
# ========================
pydantic==2.9.2
pydantic-core==2.23.4
typing-extensions>=4.8.0

# ========================
# Compatibility Fixes
# ========================
numpy==2.2.4
setuptools>=75.0.0
wheel>=0.44.0
build>=1.3.0